storage and config moved to internal directory
